+++ ton-wallet-app/src/screens/index.ts
@@ -8,3 +8,5 @@ export * from './ImportMnemonicScreen';
 export * from './MainScreen';
 export * from './ReceiveScreen';
 export * from './SendScreen';
+export * from './Settings/SettingsScreen';
+export * from './Settings/ExportScreen';